factored out package_info into its own module; exposed package ids
